requirements.txt
adjustText>=0.8
auto_mix_prep>=0.2.0
celloracle==0.12.1
dgl>=1.1.1,<1.2.0
gseapy>=1.1.3
matplotlib>=3.6.0
networkx>=3.1,<4.0.0
numpy>=1.22.4
pandas>=1.5.3
scanpy>=1.10.3
scikit_learn>=1.2.2
scipy>=1.5.3
seaborn>=0.13.2
torch>=1.12.1,<2.0.0
tqdm>=4.65.0
umap_learn>=0.5.3
ipykernel>=6.29.5

